Get simple e2e scenario working and add bridge to electron
Scenario just selects a New Project from the native
menu, fills in the form and checks that the project
is listed.
Injects a window.bridge in non-production environments
to hold convenience methods to interact with 
electron.
In the actual test a handy executeFunction has your
JS functions evaluated and executed in the test
browser